Module 8

+++ WishList/Controllers/ItemController.cs
@@ -1,22 +1,29 @@
 ﻿using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
-        public IActionResult Index()
+        public async Task<IActionResult> Index()
         {
-            var model = _context.Items.ToList();
-
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
+            var model = _context.Items.Where(i => i.User.Id == currentUser.Id).ToList();
             return View("Index", model);
         }
 
@@ -27,16 +34,20 @@ namespace WishList.Controllers
         }
 
         [HttpPost]
-        public IActionResult Create(Models.Item item)
+        public async Task<IActionResult> Create(Models.Item item)
         {
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
+            item.User = currentUser;
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
         }
 
-        public IActionResult Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            if (item.User.Id != currentUser.Id) return Unauthorized();
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
